Q&A Bot/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
langchain
langchain-community
langchain-pinecone
pinecone
PyMuPDF
tiktoken